Can now run from a DOS prompt
The Makefile now copies the relevant Cygwin files into this directory,
and modifies them all to look for `zygwin1.dll` instead of
`cygwin1.dll`.  This way we can run from a DOS prompt without an
existing Cygwin installation, and without stepping on Cygwin if it is
installed.

runme.bat is the equivalent of runme.sh.  It uses the window
configuration given in `default.view`.